Add assertions to Task child classes
There are no assertions to define assumptions about the input
parameters for the Task child public constructor

Any classes can call the constructors as they are public, but the
constructors had assumed that they are called for the right purpose.
Wrong calls would fail for the wrong errors.

Add assertions to the constructors to check if the assumptions hold.

It is better to have assertions because it will ensure the functions
input parameters follows the assumption.