Add update task functionality
No way to update a task detail unless user deletes and adds it again.

This causes inconvenience to the user. Also, since we can update
mark within the system, detail should also be allowed to update within
the system.

Let's,
- add an update command that allows user to select a specific task to
  modify the details.
- add an updateTask function to TaskList to get the specific task
- add an updateDetails function to each of the Task child classes to
  handle update for each of their own type of details field.

Splitting the implementation this way is to maintain the abstraction
and encapsulation properties of OOP programming for the classes.